Prepare v1.6: fix x64 overflows, bump version, changelog, CI
- Fix 32-bit width*height overflow in batch convert BGRA conversion
  (undersized allocation -> heap overflow for images over ~1 gigapixel);
  force 64-bit math in the same pattern in Helpers alpha compositing,
  JPEGImage alpha scan, and CHistogram sampling/means
- Bump version resources from 1.3.46.0 to 1.6.0.0 (rc + resource.h;
  the WiX installer binds its version from the exe automatically)
- Fill in the missing [1.6] changelog entries (6 commits of features
  and fixes after the first feature batch were unlisted)
- CI: stop running ctest (GitHub-hosted runners have no GPU, so GPU
  tests only skip or fail); keep compiling all test targets and add
  the missing HighwayLaneTest so registered tests stay buildable

+++ src/JPEGView/BatchConvertDlg.cpp
@@ -289,7 +289,7 @@ static void* LoadImageToBGRA(LPCTSTR sFileName, int& nWidth, int& nHeight, int& 
 
 	// Ensure BGRA 4-channel
 	if (nBPP == 3) {
-		int nSize = nWidth * nHeight * 4;
+		size_t nSize = (size_t)nWidth * nHeight * 4;
 		uint8* pBGRA = new(std::nothrow) uint8[nSize];
 		if (pBGRA == NULL) {
 			free(pPixels);
@@ -298,8 +298,8 @@ static void* LoadImageToBGRA(LPCTSTR sFileName, int& nWidth, int& nHeight, int& 
 		uint8* pSrc = (uint8*)pPixels;
 		int nSrcRowPadded = (nWidth * 3 + 3) & ~3;
 		for (int y = 0; y < nHeight; y++) {
-			uint8* pS = pSrc + y * nSrcRowPadded;
-			uint8* pD = pBGRA + y * nWidth * 4;
+			uint8* pS = pSrc + (size_t)y * nSrcRowPadded;
+			uint8* pD = pBGRA + (size_t)y * nWidth * 4;
 			for (int x = 0; x < nWidth; x++) {
 				pD[x * 4 + 0] = pS[x * 3 + 0]; // B
 				pD[x * 4 + 1] = pS[x * 3 + 1]; // G
